some changes to targets

+++ src/GksGpu/Definitions/AccumulatorDataType.h
@@ -1,7 +1,7 @@
 #ifndef AccumulatorDataType_H
 #define AccumulatorDataType_H
 
-typedef float realAccumulator;
-//typedef double realAccumulator;
+//typedef float realAccumulator;
+typedef double realAccumulator;
 
 #endif
